feat: Adds access token management
Introduces token repository and use case for handling access tokens.
This allows saving, retrieving, and clearing access tokens securely.

Implements a custom URL scheme for OAuth2 callback to receive the access token.
The SceneDelegate is updated to handle the incoming URL and store the token using the TokenRepository.

Updates the LoginView to emit an event when the Google login button is tapped.
Updates the NetworkConfiguration to empty baseUrl and appID.

feat: accessToken manage